\name{Generator of Experiments}
\alias{gerexp-package}

\docType{package}

\title{Generator of Experiments}

\description{O pacote gerexp foi elaborado com o intuito principal de gerar variáveis aleatórias de acordo com a metodologia experimental e o tipo de efeito de cada tratamento.
}

\details{Em algumas situações, temos o interesse em simular uma variável aleatória de acordo com o procedimento experimental e cujo as diferenças entre tratamentos pode ser pré-determinada. Em um delineamento inteiramente ao acaso com dois tratamentos por exemplo, podemos ter interesse em simular uma variável aleatória cujo o tratamento A terá um efeito de 1 desvio e o tratamento B um efeito de 3 desvios em relação a uma determinada média geral. 
}

\author{
Ivan Bezerra Allaman <ivanalaman@gmail.com> and José Cláudio Faria <joseclaudio.faria@gmail.com>.

Maintainer: Ivan Bezerra Allaman <ivanalaman@gmail.com>
}

\keyword{package}
