\name{gerexp-internal}
\title{Internal gerexp functions}
\alias{latin}
\alias{mgsub}
\alias{makeContrasts}
\alias{makeInteraction}
\alias{makeTreatments}
\alias{makeXBeta}
\description{Internal gerexp functions.}
\details{These are not to be called by the user and are undocumented.}
\keyword{internal}
