\name{lba-internal}
\title{Internal lba functions}
\alias{constrainAB}
\alias{identif}
\alias{rdirich}
\alias{rescaleB}
\alias{constmleEA}
\alias{constmleEB}
\alias{constmleFA}
\alias{constmleFB}
\alias{constmleFEA}
\alias{constmleFEB}
\alias{constmleFEalabama}
\alias{lba.mle.logit.A}
\alias{lba.mle.logit.AB}
\alias{lba.mle.logit.B}
\alias{logit.A}
\alias{logit.AB}
\alias{logit.B}
\description{Internal lba functions.}
\details{These are not to be called by the user and are undocumented.}
\keyword{internal}
